Add tests for both Default and Custom usage
I can't seem to get fixtures working, so I'm stuck with just creating
the Model objects as needed for testing. But I can leverage the database
schema and Model objects themselves--from the dummy Rails app in
test/dummy. This is sufficient for now, but I'd prefer to use fixtures
if I can figure this out later.