Add documentation comments within `has_statuses`
Document the inner workings of `has_status` with YARD-style comments.
However... it doesn't show up when running YARD because the comments
are all within the `has_statuses` method. Not sure what level of
meta-programming I'd need to leverage to fix this... So this is really
just for clarity of intent within the code at this point.